fix(bump): set ignoreReverted to false
Set the ignoreRevered option for `conventional-recommended-bump`
to false.

Documentation: https://github.com/conventional-changelog/conventional-recommended-bump/blob/1476f1e689b8db86b1d6b2cee3ce0cd0177df412/README.md#ignorereverted

We want to release a patch version when code changes are made as a
result of a reverted commit.

For example, submitting an enhancement, only to find out that the
enhancement breaks the tool or library. Quickest fix is to revert the
enhancement (using `git revert`) and get the library working again. I
would expect that reverting the enhancement commit would lead to a new
patch release.

Because `conventional-commits-detector` cannot detect the commit
convention used by a reverted commit, we must set a default
convention. Therefore, we set `angular` as the default convention
in the event that a convention can't be detected by
`conventional-commits-detector`.

Closes #5