chore(build): update build dependencies
Update installed dependencies so that the build uses the latest versions
of packages instead of those versions that are installed in the
`node_modules/` directory provided by the CI's cache.

This ensures the build always runs with the latest versions of build and
runtime dependencies.

Also, run `prune` to remove any packages that are not listed in
`package.json`, or are not used as a sub-dependency of a dependency
listed in `package.json`.

The `prune` command, at this time, does not work with an npm shrinkwrap
file.