refactor: remove npm publish feature
We have received numerous requests to support projects that want to
generate a GitLab release page, and associated git tag, but do not want
to publish anything to an `npm`-compatible registry.

https://gitlab.com/hyper-expanse/semantic-release-gitlab/issues/11

Reasons include private code that should only be accessed via git URLs,
or projects that are not packages, but applications, that still want the
release page functionality of `semantic-release-gitlab`, but not the
publishing feature.

BREAKING CHANGE:

We have removed support for publishing a package to an `npm`-compatible
registry. Publishing packages to an `npm`-compatible registry should be
handled using a tool like `npm-publish-git-tag`.

https://www.npmjs.com/package/npm-publish-git-tag

Starting now `semantic-release-gitlab` only handles generating a GitLab
release page for each new version, generating the git tag associated with
the release, and posting comments on issues associated with changes in
the release.

Closes #13
Closes #12
Closes #11