fix(index): support releases from a branch
Support the ability ability to run `semantic-release-gitlab` off of a
branch.

Occasionally people will encounter the following scenario:

Someone has released a new major version of their project. A consumer
of that project reports a bug in the earlier major version, and can't,
for whatever reason, upgrade to the latest major version at this time.

That consumer would greatly benefit if the project could quickly submit
a patch against the earlier major version and have
`semantic-release-gitlab` automatically release that version to GitLab.

The owner of the project should be able to create a dedicated branch off
of the latest code for the previous major version, push a bug fix to
that branch, and have `semantic-release-gitlab` automatically release a
new patch version.

Fixes #15