fix: commenting for GitLab >11
Upgrade dependency that posts comments to issues and merge requests so
that it will work with GitLab version 11 and greater. The dependency
switched to using GitLab API version 4, which is now the only version
supported. Version 3 of their API was removed several weeks ago with
the release of GitLab version 11.

BREAKING CHANGE:

Requires a version of GitLab with API version 4, and it must be
accessible.