fix(index): delete git tag on failure
`semantic-release-gitlab` creates a git tag on disk to be used by
`conventional-gitlab-releaser` to generate a changelog. If there's a
failure during the release process (such as creating a GitLab Release,
or posting comments to GitLab) then the git tag is left on disk, which
can, and has, caused problems when re-running the release tool.

When the release tool is re-run within the same folder, `semantic-release-gitlab`
will see the previously created git tag. `semantic-release-gitlab` will
assume that the release has already happened, and will report that there
is nothing to release.

Therefore, delete the the git tag if `semantic-release-gitlab` encounters
an error during the release process.

Fixes #35