feat(index): remove major version zero support
BREAKING CHANGE:

Remove support for incrementing a major version zero, a semantic
version that starts with the number 0, such as `0.1.0`, in which a
feature or bug fix is relased as a PATCH, the third position
in the number, instead of the feature being released as a MINOR,
the second, or middle, position in the number. Major version
zero support guaranteed that a project on major version zero would
remain on major version zero unless the maintainer manually bumped
the version up to at least version `1`.

After maintianing major version zero support, I feel major version zero
support has led project maintainers to get into a rut in which they
continue to release major version zero releases, never incrementing
to a formal version 1, or higher.